+++ src/ledger/LedgerGateway.h
@@ -34,7 +34,7 @@ class LedgerGateway
 
     virtual uint64_t getLedgerNum() = 0;
     virtual uint64_t getCloseTime() = 0;
-    virtual int32_t getTxFee() = 0;
+    virtual int64_t getTxFee() = 0;
 };
 }
